draft/requirements_old.txt
librosa
numpy==1.20
pandas
praat-parselmouth
pyannote.audio==1.1.2
pychord
tensorboard
torch
torchvision
